+++ CHANGELOG.md
@@ -5,3 +5,14 @@
 ### Added
 
 - Initial Release of View Ball Fix
+
+## [1.0.1] - 2025-07-27
+
+### Added
+
+- Fixed some performance issues
+- Calibration is more reliable now
+- Calibration cleanup is more reliable now
+- Added a new button to fix the calibration window if Unity crashes while calibrating
+- Updated the calibration message in Play Mode, after I found an issue
+- Updated console debug messages